Removing unnecessary files and changed some tests
